requirements.txt
networkx
anyio
fastapi
uvicorn
websockets
python-multipart
google-genai
